install stripe

+++ frontend/src/services/cart.js
@@ -9,25 +9,31 @@ import useAuth from '../stores/Auth';
 export default function useCart() {
 
     function addCart(productId){
-        useAxios.get('/users/cart/add/' + 2 + '/' + productId)
-            .then(response => {
-                useWishlist().deleteWishlist(productId)
-                useRefresh().refreshCartItemsNumber = !useRefresh().refreshCartItemsNumber
-                useAlert().topAlert('success', 'Successfully added to cart')
+        if (useAuth().isAuthenticated) {
+            useAxios.get('/users/cart/add/' + useAuth().user.id + '/' + productId, {
+                headers: { Authorization: 'Bearer ' + useToken().getToken() }
             })
-            .catch(error => {
-                console.log(error);
-                useAlert().topAlert('error', error.response.data.message, 'bottom-end')
-            });
+                .then(response => {
+                    useWishlist().deleteWishlist(productId)
+                    useRefresh().refreshCartItemsNumber = !useRefresh().refreshCartItemsNumber
+                    useAlert().topAlert('success', 'Successfully added to cart')
+                })
+                .catch(error => {
+                    console.log(error);
+                    useAlert().topAlert('error', error.response.data.message, 'bottom-end')
+                });
+        }else{
+            useAlert().topAlert('error', 'Please login first...', 'bottom-end')
+        }
     }
 
-    function  getCartItems() {
-        return useAxios.get('/users/cart/' + 2, {
+    async function  getCartItems() {
+        return await useAxios.get('/users/cart/' + useAuth().user.id, {
             headers: { Authorization: 'Bearer ' + useToken().getToken() }
         })
             .then(response => {
-                return response.data;
                 if (response.data.carts.length === 0) useAlert().centerDialogAlert('info', 'Your cart list is empty')
+                return response.data;
             })
     }
